==> notebooks/02_analyzing_optimization.ipynb <==
import json
import os

import pandas as pd

path = '../results/02_bayesian_optimization'

# 1. Bayes Optimization Results

def parse_json_file(file_path: str) -> dict:
    """Parse a json file and return its content as a dictionary."""
    with open(file_path, 'r') as file:
        return json.load(file)

def get_results_from_trial(
    trial_path: str, 
    batch_size: int, 
    trial_num: int
) -> dict:
    """Get the results from a trial and return them as a dictionary."""
    json_file = next(f for f in os.listdir(trial_path) if f.endswith('.json'))
    curr_result = parse_json_file(os.path.join(trial_path, json_file))
    
    result = curr_result['hyperparameters']['values']
    result['epochs'] = curr_result['best_step']
    result['loss'] = curr_result['score']
    result['batch_size'] = batch_size
    result['trial'] = trial_num
    
    return result

def results_parser(path: str) -> pd.DataFrame:
    """Parse the results from the trials and return them as a DataFrame."""
    results_df = pd.DataFrame()

    for rd in os.listdir(path):
        curr_path = os.path.join(path, rd)
        batch_size = int(rd.split('_')[-1])
        
        trials = [
            trial 
            for trial in os.listdir(curr_path) 
            if not trial.endswith('.json')
        ]

        for trial in trials:
            trial_num = int(trial.split('_')[1])
            curr_trial = os.path.join(curr_path, trial)
            result = get_results_from_trial(curr_trial, batch_size, trial_num)
            
            results_df = pd.concat(
                [results_df, pd.DataFrame(result, index=[0])], 
                ignore_index=True
            )
    
    return results_df.sort_values('loss')


## 1.1 Pretrain BERT

### 1.1.1 Book Reviews

# top 10 melhores resultados (bert com book reviews)
results_lit_review = results_parser(f'{path}/bert/book_reviews')
results_lit_review.head(10)

# média de resultado por batch_size e arquitetura (bert com book reviews)
results_lit_review.groupby(
    ['batch_size']
).agg(
    {'loss': ['mean', 'std', 'median']}
).sort_values(['batch_size'])

### 1.1.2 TV

# top 10 melhores resultados (bert com tv reviews)
results_tv_review = results_parser(f'{path}/bert/tv_reviews')
results_tv_review.head(10)

# média de resultado por batch_size e arquitetura (bert com tv reviews)
results_tv_review.groupby(
    ['batch_size']
).agg(
    {'loss': ['mean', 'std', 'median']}
).sort_values(['batch_size'])

## 1.2 Fine Tuning

def read_json(path: str) -> dict:
    """Reads a json file and returns a dictionary."""
    with open(path, "r") as f:
        return json.load(f)


def best_hyperparams(results: dict) -> dict:
    """Returns the best hyperparameters based on the F1 score."""
    best_values = {
        'trial': None,
        'F1': 0,
    }
    for trial, result in results.items():
        if result['value'] > best_values['F1']:
            best_values['trial'] = trial
            best_values['F1'] = result['value']

    print(f'Best trial: {best_values["trial"]}.')

    return results[best_values['trial']]

### 1.2.2 TV

# procurando qual tentativa apresentou o melhor desempenho
tv_results1 = read_json(f'{path}/fine_tuning/tv/baseline.json')
best_hyperparams(tv_results1)

# procurando qual tentativa apresentou o melhor desempenho
tv_results2 = read_json(f'{path}/fine_tuning/tv/bert_visio.json')
best_hyperparams(tv_results2)

# procurando qual tentativa apresentou o melhor desempenho
tv_results3 = read_json(f'{path}/fine_tuning/tv/d_bert_visio.json')
best_hyperparams(tv_results3)

# procurando qual tentativa apresentou o melhor desempenho
tv_results4 = read_json(f'{path}/fine_tuning/tv/d_bert_visio_ffn.json')
best_hyperparams(tv_results4)

# procurando qual tentativa apresentou o melhor desempenho
tv_results5 = read_json(f'{path}/fine_tuning/tv/lora_baseline.json')
best_trial = (None, 0.0)
for trial, infos in tv_results5.items():
    if 'metrics' in infos.keys():
        if infos['metrics']['f1'] > best_trial[1]:
            best_trial = (trial, infos['metrics']['f1'])
tv_results5[best_trial[0]]

### 1.2.3 ReLi

# procurando qual tentativa apresentou o melhor desempenho
reli_results1 = read_json(f'{path}/fine_tuning/reli/baseline.json')
best_hyperparams(reli_results1)

# procurando qual tentativa apresentou o melhor desempenho
reli_results2 = read_json(f'{path}/fine_tuning/reli/litera_bert.json')
best_hyperparams(reli_results2)

# procurando qual tentativa apresentou o melhor desempenho
reli_results3 = read_json(f'{path}/fine_tuning/reli/d_litera_bert.json')
best_hyperparams(reli_results3)

# procurando qual tentativa apresentou o melhor desempenho
reli_results4 = read_json(f'{path}/fine_tuning/reli/d_litera_bert_ffn.json')
best_hyperparams(reli_results4)

# procurando qual tentativa apresentou o melhor desempenho
reli_results5 = read_json(f'{path}/fine_tuning/reli/lora_baselie.json')
best_trial = (None, 0.0)
for trial, infos in reli_results5.items():
    if 'metrics' in infos.keys():
        if infos['metrics']['f1'] > best_trial[1]:
            best_trial = (trial, infos['metrics']['f1'])
reli_results5[best_trial[0]]
