==> notebooks/06_gpt_results.ipynb <==
from ast import literal_eval

import pandas as pd

results = (
    pd.read_csv('../datasets/gpt/tv/test.csv')
    .rename(
        columns={'explicit aspects': 'aspects'}
    )
)

gpt_results = results.columns[2:]

for col in ['aspects'] + gpt_results.tolist():
    results[col] = results[col].apply(literal_eval)

results.head()

# se está em GPT_RESULTS e está em ASPECTS: TP
def true_positive(y_true, y_pred):
    tp = 0
    y_true = set(y_true)
    y_pred = set(y_pred)
    for each in y_pred:
        if each in y_true:
            tp += 1
    return tp

# se está em GPT_RESULTS e NÃO está em ASPECTS: FP
def false_positive(y_true, y_pred):
    fp = 0
    y_true = set(y_true)
    y_pred = set(y_pred)
    for each in y_pred:
        if each not in y_true:
            fp += 1
    return fp

# se está em ASPECTS e NÃO está em GPT_RESULTS: FN
def false_negative(y_true, y_pred):
    fn = 0
    y_true = set(y_true)
    y_pred = set(y_pred)
    for each in y_true:
        if each not in y_pred:
            fn += 1
    return fn

def compute(data, by, func):
    return data.apply(lambda x: func(x['aspects'], x['curie_results']))

def precision(tp, fp):
    return tp / (tp + fp)

def recall(tp, fn):
    return tp / (tp + fn)

def f1_score(precision, recall):
    return 2 * (precision * recall) / (precision + recall)

for gpt in gpt_results:

    tp, fp, fn = 0, 0, 0
    for _, row in results.iterrows():
        y_true, y_pred = row['aspects'], row[gpt]
        tp += true_positive(y_true, y_pred)
        fp += false_positive(y_true, y_pred)
        fn += false_negative(y_true, y_pred)

    p = precision(tp, fp)
    r = recall(tp, fn)
    f1 = f1_score(p, r)

    print(f'{gpt} -> p = {p:.4f}, r = {r:.4f}, f1 = {f1:.4f}')
